+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - `add_note()` to generate a note with a sum of harmonics.
     - `add_major_chord()`
     - `add_minor_chord()`
-
+- demos module:
+    - `demo1()`
 
 ## [Forsynth 0.1 "Stockhausen"] 2021-02-19